docs: Refactor sandpack usage to account for react examples (NOTE: a bit hacky). Update react docs mdx.
